Extended the maximum range of number able to print
Doubled the range of the function from overflowing after about 50 numbers to at least over 100

+++ Fibonacci series.c
@@ -1,11 +1,11 @@
 #include<stdio.h>    
 void printFibonacci(int n){    
-    static int n1=0,n2=1,n3;    
+    static unsigned long n1=0,n2=1,n3;    
     if(n>0){    
          n3 = n1 + n2;    
          n1 = n2;    
          n2 = n3;    
-         printf("%d ",n3);    
+         printf("%lu ",n3);    
          printFibonacci(n-1);    
     }    
 }
